requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
discord.py==2.3.2
distro==1.7.0
distro-info==1.1+ubuntu0.1
frozenlist==1.4.1
git-filter-repo==2.34.0
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
jeepney==0.7.1
jellyfish==0.11.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
oauthlib==3.2.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
systemd-python==234
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
vboxapi==1.0
wadllib==1.3.6
yarl==1.9.4
zipp==1.0.0
